+++ src/mergeTiles/fixBorderMismatch.js
@@ -1,4 +1,4 @@
-const { genArray, arePointsEqual } = require("../utils");
+const { genArray, arePointsEqual } = require("../utils/utils");
 const {
   getNextPointByIdx,
 } = require("../utils/pointArrangement/getPointFromList");
